requirements.txt
tensorflow==2.4.1
streamlit==0.73.0
opencv-contrib-python-headless==4.4.0.46
numpy==1.20.1
Pillow==7.0.0
h5py<3.0.0
